Move `EventMiddlewareMock` to separate file
