+++ Sources_v3/WebSocketClient/WebSocketClient_Tests.swift
@@ -525,19 +525,6 @@ extension WebSocketEngineError: Equatable {
     }
 }
 
-/// A test middleware that can be initiated with a closure/
-final class EventMiddlewareMock: EventMiddleware {
-    var closure: (Event, @escaping (Event?) -> Void) -> Void
-    
-    init(closure: @escaping (Event, @escaping (Event?) -> Void) -> Void = { $1($0) }) {
-        self.closure = closure
-    }
-    
-    func handle(event: Event, completion: @escaping (Event?) -> Void) {
-        closure(event, completion)
-    }
-}
-
 class MockBackgroundTaskScheduler: BackgroundTaskScheduler {
     var beginBackgroundTask_called: Bool = false
     var beginBackgroundTask_expirationHandler: (() -> Void)?
